Implement a basic timer system for tracking game times
The timer is configurable to track time Downcounting or Upcounting
and has some other features like being able to adjust the player
time for each completed turn.

The implementation is currently completely polling based (i.e. you
need to check the time if you want an update) and would probably
benefit from some asynchronicity but this implementation should be
perfectly fine for now.